Update README.md
A multiplex cinema chain wants to run a local campaign to encourage more visitors to its cinemas and to increase sales of popcorn, soft drinks, sweets and mineral water. This chain has 4 screening rooms (including a VIP room) in a shopping center located in a prime area and 5 rooms (including an IMAX room) in a shopping complex frequented by the middle class population.

In this project, 12 months of transactional data from products sold to customers who are part of the chain's loyalty club will be analyzed and clustered using K-Means, so that the company can target its marketing campaigns and promote online co-sales (along with ticket sales) for customers to pick up the products when they arrive at the cinema.